Skip collecting app versions for apps in the org-* namespaces
The `app-operator.giantswarm.io/version` label is not mandatory there (CAPI clusters)
so we do not know the version of the app-operator reconciling these apps. We could check
the app-operator in the same, org namespace but then the validation after that that this
part does makes little sense as if we were able to get the version than the opp-operator
obviously does exist.